+++ apps/block_scout_web/lib/block_scout_web/views/bridged_tokens_view.ex
@@ -5,7 +5,7 @@ defmodule BlockScoutWeb.BridgedTokensView do
 
   alias BlockScoutWeb.ChainView
   alias Explorer.Chain
-  alias Explorer.Chain.{BridgedToken, Token}
+  alias Explorer.Chain.{Address, BridgedToken, Token}
 
   @owl_token_amb "0x0905Ab807F8FD040255F0cF8fa14756c1D824931"
   @owl_token_omni "0x750eCf8c11867Ce5Dbc556592c5bb1E0C6d16538"